backend/requirements.txt
fastapi
python-dotenv
uvicorn
pydantic
python-multipart